+++ fme_packager/cli.py
@@ -3,13 +3,17 @@ fme_packager command line interface.
 """
 
 import inspect
+import json as json_lib
 import os
 import shutil
+import subprocess
 import sys
 import sysconfig
 from pathlib import Path
 
 import click
+from click import BadParameter
+from cookiecutter.exceptions import OutputDirExistsException, FailedHookException
 from cookiecutter.main import cookiecutter
 
 import fme_packager
@@ -43,7 +47,11 @@ def init(template):
     TEMPLATE -- Name of the template to use.
     """
     print("Enter the values to use for the template. Press Enter to accept the [default].")
-    cookiecutter(COOKIECUTTER_TEMPLATES[template])
+    try:
+        cookiecutter(COOKIECUTTER_TEMPLATES[template])
+    except (OutputDirExistsException, FailedHookException) as e:
+        print(e)  # Don't show stack trace for validation errors
+        sys.exit(1)
 
 
 @cli.command()
@@ -97,7 +105,7 @@ def verify(file, verbose, json):
 
 
 @cli.command()
-@click.argument("file", type=click.Path(exists=True, file_okay=True, dir_okay=False))
+@click.argument("file", type=click.Path(exists=True))
 def summarize(file):
     """
     Extract a JSON representation of an .fpkg file.
@@ -105,7 +113,7 @@ def summarize(file):
     FILE -- Path to an FME .fpkg package file.
     """
     result = summarize_fpkg(file)
-    print(result)
+    print(json_lib.dumps(result, indent=2))
 
 
 @cli.command()
@@ -114,7 +122,7 @@ def summarize(file):
     prompt=True,
     default=lambda: os.environ.get("FME_HOME"),
     show_default="FME_HOME environment variable",
-    type=click.Path(exists=True, file_okay=False, writable=True),
+    type=click.Path(exists=True, file_okay=False, writable=False),
     required=True,
     help="Path to FME installation",
 )
@@ -126,37 +134,74 @@ def summarize(file):
     required=True,
     help="Path to Python environment's site-packages",
 )
-def config_env(fme_home, site_packages_dir):
+@click.option(
+    "--verify/--no-verify",
+    is_flag=True,
+    default=True,
+    help="Skip verification of fmeobjects access",
+)
+def config_env(fme_home, site_packages_dir, verify):
     """
     Configure a Python environment for access to fmeobjects
     and the Python libraries included with FME.
     """
+    fme_home = Path(fme_home)
+    if not (fme_home / "fme").is_file() and not (fme_home / "fme.exe").is_file():
+        raise BadParameter(f"'{fme_home}' does not contain an FME executable")
+
     leaf_dir = "python%s%s" % (sys.version_info.major, sys.version_info.minor)
     paths_to_add = [
-        "#" + fme_home,
-        os.path.join(fme_home, "python"),
-        os.path.join(
-            fme_home,
-            "python",
-            f"fme-plugins-py{sys.version_info.major}{sys.version_info.minor}.zip",
-        ),
-        os.path.join(fme_home, "python", leaf_dir),
-        os.path.join(fme_home, "fmeobjects", leaf_dir),
+        "#" + fme_home.as_posix(),  # fme_env.py reads this line
+        (fme_home / "python").as_posix(),  # fmeobjects API and third-party libraries
+        (fme_home / "python" / leaf_dir).as_posix(),  # third-party binary libraries
     ]
 
-    dst_pth = os.path.join(site_packages_dir, "fme_env.pth")
-    print("Writing " + dst_pth)
-    with open(dst_pth, "w") as f:
+    # Internal libraries bundle on Windows since FME 2023.2 (FMEENGINE-78545)
+    # Mac/Linux since FME 2025.1 (FOUNDATION-22)
+    internal_libs_zip = (
+        fme_home / "python" / f"fme-plugins-py{sys.version_info.major}{sys.version_info.minor}.zip"
+    )
+    if internal_libs_zip.is_file():
+        paths_to_add.append(internal_libs_zip.as_posix())
+
+    # Location of fmeobjects before FME 2023.2 (FMEENGINE-78545)
+    legacy_fmeobjects_path = fme_home / "fmeobjects" / leaf_dir
+    if legacy_fmeobjects_path.is_dir():
+        paths_to_add.append(legacy_fmeobjects_path.as_posix())
+
+    site_packages_dir = Path(site_packages_dir)
+    dst_pth = site_packages_dir / "fme_env.pth"
+    print(f"Writing {dst_pth}")
+    with dst_pth.open("w") as f:
         for path in paths_to_add:
             f.write(path + "\n")
         f.write("import fme_env\n")
 
-    dst_py = os.path.join(site_packages_dir, "fme_env.py")
-    print("Writing " + dst_py)
+    dst_py = site_packages_dir / "fme_env.py"
+    print(f"Writing {dst_py}")
     src = Path(inspect.getfile(fme_packager)).parent / "fme_env.py"
     shutil.copy(src, dst_py)
 
-    print("\nThis Python environment is now set up for access to FME and fmeobjects.")
+    if verify:
+        print("Verifying access to fmeobjects...")
+        for test_import in [
+            "from fmeobjects import FMEFeature",
+            "import pluginbuilder",
+            "import fmewebservices",
+        ]:
+            try:
+                subprocess.run(
+                    [sys.executable, "-c", f"import sys;sys.tracebacklimit=0;{test_import}"],
+                    check=True,
+                    capture_output=True,
+                )
+            except subprocess.CalledProcessError as e:
+                print(f"{test_import}: {e.stderr.decode().strip()}")
+                print("Environment configuration failed.")
+                print(f"Using Python {sys.version} at {sys.executable}")
+                return sys.exit(1)
+
+    print("\nThis Python environment is now set up for access to fmeobjects.")
     print("If the FME install location changes, re-run this script to update paths.")