+++ fme_packager/metadata.py
@@ -1,27 +1,28 @@
+from importlib.resources import open_text
 import json
 import os
+from pathlib import Path
+from typing import Union
 
 from ruamel.yaml import YAML
 
 
-def load_fpkg_metadata(fpkg_path):
+def load_fpkg_metadata(fpkg_path: Union[str, os.PathLike]):
     """
     Load package.yml as a FMEPackageMetadata object.
 
-    :param str fpkg_path: Path to the directory where package.yml exists.
+    :param fpkg_path: Path to the directory where package.yml exists.
     :rtype: FMEPackageMetadata
     """
-    with open(os.path.join(fpkg_path, "package.yml")) as f:
+    with open(Path(fpkg_path) / "package.yml") as f:
         return FMEPackageMetadata(YAML(typ="safe").load(f))
 
 
-def load_metadata_json_schema():
+def load_metadata_json_schema() -> dict:
     """
     Load in package metadata specification and requirements.
-
-    :rtype: dict
     """
-    with open(os.path.join(os.path.dirname(__file__), "spec.json")) as f:
+    with open_text("fme_packager", "spec.json") as f:
         return json.load(f)